+++ checkpoint/orbax/checkpoint/experimental/v1/_src/handlers/pytree_handler_test_base.py
@@ -20,11 +20,12 @@ from __future__ import annotations
 
 import asyncio
 import contextlib
+import dataclasses
 import datetime
 import functools
 import json
 import threading
-from typing import Any, Awaitable, Iterator, List, Sequence
+from typing import Any, Awaitable, Iterator, List, Sequence, Type
 from unittest import mock
 
 from absl.testing import parameterized
@@ -87,6 +88,70 @@ Path = path_types.Path
 
 async def _run_awaitable(awaitable: Awaitable[Any]) -> Any:
   return await awaitable
+
+
+# Custom dataclasses for testing custom leaf handlers.  PyType check requires
+# these defines outside of the test.
+@dataclasses.dataclass
+class Point:
+  x: int
+  y: float
+
+
+@dataclasses.dataclass
+class AbstractPoint:
+  x: Type[int] = int
+  y: Type[float] = float
+
+
+class PointLeafHandler(serialization_types.LeafHandler[Point, AbstractPoint]):
+  """A custom leaf handler for testing."""
+
+  def __init__(self, context: context_lib.Context | None = None):
+    del context
+
+  async def serialize(
+      self,
+      params: Sequence[serialization_types.SerializationParam[Point]],
+      serialization_context: serialization_types.SerializationContext,
+  ) -> Awaitable[None]:
+
+    def _background_serialize():
+      if multihost.is_primary_host(0):
+        for param in params:
+          with open(
+              serialization_context.parent_dir.path / f'{param.name}.txt',
+              'w',
+          ) as f:
+            f.write(json.dumps(dataclasses.asdict(param.value)))
+
+    return asyncio.to_thread(_background_serialize)
+
+  async def deserialize(
+      self,
+      params: Sequence[serialization_types.DeserializationParam[AbstractPoint]],
+      deserialization_context: serialization_types.DeserializationContext,
+  ) -> Awaitable[Sequence[Point]]:
+
+    def _deserialize_impl():
+      ret = []
+      for param in params:
+        with open(
+            deserialization_context.parent_dir / f'{param.name}.txt',
+            'r',
+        ) as f:
+          ret.append(Point(**json.loads(f.read())))
+
+      return ret
+
+    return asyncio.to_thread(_deserialize_impl)
+
+  async def metadata(
+      self,
+      params: Sequence[serialization_types.DeserializationParam[None]],
+      deserialization_context: serialization_types.DeserializationContext,
+  ) -> Sequence[AbstractPoint]:
+    return [AbstractPoint()] * len(params)
 
 
 class PyTreeHandler:
@@ -351,6 +416,8 @@ class PyTreeHandlerTestBase:
           return str
         if isinstance(value, optax.EmptyState):
           return None
+        if isinstance(value, Point):
+          return AbstractPoint()
         raise ValueError(f'Unrecognized type: {type(value)}.')
 
       expected_metadata = jax.tree.map(
@@ -1351,6 +1418,9 @@ class PyTreeHandlerTestBase:
       class PlusOneHandler(scalar_leaf_handler.ScalarLeafHandler):
         """A custom handler that adds one to all scalar values."""
 
+        def __init__(self, context: context_lib.Context | None = None):
+          super().__init__(context=context)
+
         async def serialize(
             self,
             params: Sequence[scalar_leaf_handler.ScalarSerializationParam],
@@ -1366,17 +1436,15 @@ class PyTreeHandlerTestBase:
           return await super().serialize(updated_params, serialization_context)
 
       leaf_registry = registry.BaseLeafHandlerRegistry()
-      leaf_registry.add(int, int, PlusOneHandler())
+      leaf_registry.add(int, int, PlusOneHandler)
 
       with handler_with_options(
           leaf_handler_registry=leaf_registry,
           array_metadata_store=None,
           use_zarr3=True,
       ) as handler:
-        # TODO(b/430598877) V1 Registry error message should be returned intead.
-        with self.assertRaisesRegex(
-            ValueError, "TypeHandler lookup failed for: type=<class 'float'>"
-        ):
+        # TODO(b/430598877) Return V1 Registry error message.
+        with self.assertRaisesRegex(ValueError, 'TypeHandler lookup failed'):
           handler.save(self.directory, {'a': 3, 'b': 1.0})
 
         handler.save(self.directory, {'a': 3})
@@ -2056,3 +2124,47 @@ class PyTreeHandlerTestBase:
               r"Missing 1 keys in structure path \(\), including: \['z'\]",
           ):
             restore_handler.load(directory, reference_item)
+
+    @parameterized.product(use_zarr3=(True, False), use_ocdbt=(True, False))
+    def test_custom_leaf_handler(self, use_zarr3: bool, use_ocdbt: bool):
+
+      pytree = {
+          'point1': Point(1, 2),
+          'point2': Point(3, 4),
+          'nested': {
+              'point3': Point(5, 6),
+              'point4': Point(7, 8),
+          },
+          'string_leaf': 'string_leaf',
+          'number': 123,
+          'pytree': self.pytree,
+      }
+
+      array_metadata_store = ARRAY_METADATA_STORE
+
+      leaf_handler_registry = registry.StandardLeafHandlerRegistry()
+      leaf_handler_registry.add(Point, AbstractPoint, PointLeafHandler)
+
+      def _as_abstract_type(x):
+        if isinstance(x, Point):
+          return AbstractPoint
+        return as_abstract_type(x)
+
+      with handler_with_options(
+          use_ocdbt=use_ocdbt,
+          leaf_handler_registry=leaf_handler_registry,
+          array_metadata_store=array_metadata_store,
+          use_zarr3=use_zarr3,
+      ) as checkpoint_handler:
+        checkpoint_handler.save(self.directory, pytree)
+        abstract_pytree = jax.tree.map(_as_abstract_type, pytree)
+        restored = checkpoint_handler.load(self.directory, abstract_pytree)
+
+        test_utils.assert_tree_equal(self, pytree, restored)
+
+        self.validate_metadata(
+            expected_reference_metadata_tree=pytree,
+            actual_metadata=checkpoint_handler.metadata(self.directory),
+            pytree_metadata_options=self.pytree_metadata_options,
+            array_metadata_store=array_metadata_store,
+        )