+++ checkpoint/orbax/checkpoint/experimental/v1/_src/serialization/registry.py
@@ -19,7 +19,6 @@ from typing import Any, Dict, Sequence, Tuple, Type
 from absl import logging
 import jax
 import numpy as np
-from orbax.checkpoint.experimental.v1._src.context import context as context_lib
 from orbax.checkpoint.experimental.v1._src.serialization import array_leaf_handler
 from orbax.checkpoint.experimental.v1._src.serialization import numpy_leaf_handler
 from orbax.checkpoint.experimental.v1._src.serialization import protocol_utils
@@ -27,6 +26,7 @@ from orbax.checkpoint.experimental.v1._src.serialization import scalar_leaf_hand
 from orbax.checkpoint.experimental.v1._src.serialization import string_leaf_handler
 from orbax.checkpoint.experimental.v1._src.serialization import types
 import typing_extensions
+
 
 # The standard type and abstract type to handler mapping.
 # The type to abstract type pairs are well defined standard and users should
@@ -62,131 +62,142 @@ STANDARD_TYPE_AND_ABSTRACT_TYPE_TO_HANDLER = {
 class BaseLeafHandlerRegistry:
   """Base Leaf Handler Registry that implements the LeafHandlerRegistry Protocol."""
 
-  def __init__(self, context: context_lib.Context | None = None):
-    self._type_registry: Dict[Type[Any], types.LeafHandler[Any, Any]] = {}
+  def __init__(self):
+    self._leaf_type_registry: Dict[
+        Type[Any], Type[types.LeafHandler[Any, Any]]
+    ] = {}
     self._abstract_type_registry: Dict[
-        Type[Any], types.LeafHandler[Any, Any]
+        Type[Any], Type[types.LeafHandler[Any, Any]]
     ] = {}
 
     # for easy look up for replacement
     self._handler_to_types: Dict[
-        types.LeafHandler[Any, Any], Tuple[Type[Any], Type[Any]]
+        Type[types.LeafHandler[Any, Any]], Tuple[Type[Any], Type[Any]]
     ] = {}
-    self._context = context_lib.get_context(context)
 
   def _try_get(
-      self, ty: Type[types.Leaf]
-  ) -> types.LeafHandler[types.Leaf, Any] | None:
+      self, leaf_type: Type[types.Leaf]
+  ) -> Type[types.LeafHandler[types.Leaf, Any]] | None:
     """Returns the handler registered for a given type, if available."""
-    for registered_ty, handler in self._type_registry.items():
-      if issubclass(ty, registered_ty):
-        return handler
+    for registered_ty, handler_type in self._leaf_type_registry.items():
+      if issubclass(leaf_type, registered_ty):
+        return handler_type
 
     # no handler found
     return None
 
   def get(
-      self, ty: Type[types.Leaf]
-  ) -> types.LeafHandler[types.Leaf, Any] | None:
-    if (handler := self._try_get(ty)) is None:
+      self, leaf_type: Type[types.Leaf]
+  ) -> Type[types.LeafHandler[types.Leaf, Any]] | None:
+    if (handler_type := self._try_get(leaf_type)) is None:
       raise ValueError(
-          f'Unknown Leaf type: "{ty}". Must register it with'
+          f'Unknown Leaf type: "{leaf_type}". Must register it with'
           ' LeafHandlerRegistry.'
       )
 
-    return handler
+    return handler_type
 
   def _try_get_abstract(
-      self, abstract_ty: Type[types.AbstractLeaf]
-  ) -> types.LeafHandler[Any, types.AbstractLeaf] | None:
+      self,
+      abstract_type: Type[types.AbstractLeaf],
+  ) -> Type[types.LeafHandler[Any, types.AbstractLeaf]] | None:
     """Returns the handler registered for a given abstract type, if available."""
-    for registered_abstract_ty, handler in self._abstract_type_registry.items():
+    for (
+        registered_abstract_ty,
+        handler_type,
+    ) in self._abstract_type_registry.items():
       if typing_extensions.is_protocol(registered_abstract_ty):  # pytype: disable=not-supported-yet
         if protocol_utils.is_subclass_protocol(
-            cls=abstract_ty, protocol=registered_abstract_ty
+            cls=abstract_type, protocol=registered_abstract_ty
         ):
-          return handler
-      elif issubclass(abstract_ty, registered_abstract_ty):
-        return handler
+          return handler_type
+      elif issubclass(abstract_type, registered_abstract_ty):
+        return handler_type
 
     # no handler found
     return None
 
   def get_abstract(
-      self, abstract_ty: Type[types.AbstractLeaf]
-  ) -> types.LeafHandler[Any, types.AbstractLeaf]:
-    if (handler := self._try_get_abstract(abstract_ty)) is None:
+      self,
+      abstract_type: Type[types.AbstractLeaf],
+  ) -> Type[types.LeafHandler[Any, types.AbstractLeaf]]:
+    if (handler_type := self._try_get_abstract(abstract_type)) is None:
       raise ValueError(
-          f'Unknown AbstractLeaf type: "{abstract_ty}". Must register it with'
+          f'Unknown AbstractLeaf type: "{abstract_type}". Must register it with'
           ' LeafHandlerRegistry.'
       )
 
-    return handler
+    return handler_type
 
   def get_all(
       self,
   ) -> Sequence[types.LeafHandlerRegistryItem]:
     """Returns all registered handlers."""
-    ret = []
-
-    for (ty, handler), abstract_ty in zip(
-        self._type_registry.items(), self._abstract_type_registry
-    ):
-      ret.append((ty, abstract_ty, handler))
-    return ret
+    return [
+        (
+            leaf_type,
+            abstract_type,
+            handler_type,
+        )
+        for (leaf_type, handler_type), abstract_type in zip(
+            self._leaf_type_registry.items(), self._abstract_type_registry
+        )
+    ]
 
   def add(
       self,
-      ty: Type[types.Leaf],
-      abstract_ty: Type[types.AbstractLeaf],
-      handler: types.LeafHandler[types.Leaf, types.AbstractLeaf],
+      leaf_type: Type[types.Leaf],
+      abstract_type: Type[types.AbstractLeaf],
+      handler_type: Type[types.LeafHandler[types.Leaf, types.AbstractLeaf]],
       override: bool = False,
   ):
-    """Adds a handler for a given type and abstract type pair."""
-    current_handler = self._try_get(ty)
-    handler_abstract_ty = self._try_get_abstract(abstract_ty)
+    """Adds a handler_type for a given leaf_type and abstract_type pair."""
+    current_handler_type = self._try_get(leaf_type)
+    current_abstract_handle_type = self._try_get_abstract(abstract_type)
 
-    if not override and (current_handler or handler_abstract_ty):
+    if not override and (current_handler_type or current_abstract_handle_type):
       raise ValueError(
-          f'Type[{ty}] or Abstract_ty[{abstract_ty}] has already registered.'
-          f' current_handler: {current_handler}, handler_abstract_ty:'
-          f' {handler_abstract_ty}'
+          f'Leaf_type[{leaf_type}] or abstract_type[{abstract_type}] has'
+          f' already registered, current_handler: {current_handler_type}, '
+          f'current_abstract_handle_type: {current_abstract_handle_type}'
       )
 
     logging.vlog(
         1,
-        'add: ty[%s], abstract_ty[%s], current_handler[%s],'
-        ' handler_abstract_ty[%s]',
-        ty,
-        abstract_ty,
-        current_handler,
-        handler_abstract_ty,
+        'add: leaf_type[%s], abstract_type[%s], handler_type[%s],'
+        ' current_handler[%s], current_abstract_handle_type[%s]',
+        leaf_type,
+        abstract_type,
+        handler_type,
+        current_handler_type,
+        current_abstract_handle_type,
     )
 
-    if current_handler and (
-        handler_abstract_ty and current_handler != handler_abstract_ty
+    if current_handler_type and (
+        current_abstract_handle_type
+        and current_handler_type != current_abstract_handle_type
     ):
       raise ValueError(
-          f'Abstract_ty[{abstract_ty}] has already registered with a'
+          f'Abstract_type[{abstract_type}] has already registered with a'
           ' different type.'
       )
-    elif current_handler and not handler_abstract_ty:
+    elif current_handler_type and not current_abstract_handle_type:
       # need to remove the previous abstract type
-      _, old_abstract_ty = self._handler_to_types.pop(current_handler)
+      _, old_abstract_ty = self._handler_to_types.pop(current_handler_type)
       self._abstract_type_registry.pop(old_abstract_ty)
 
     # new type and abstract type pair
-    self._type_registry[ty] = handler
-    self._abstract_type_registry[abstract_ty] = handler
-    self._handler_to_types[handler] = (ty, abstract_ty)
+    self._leaf_type_registry[leaf_type] = handler_type
+    self._abstract_type_registry[abstract_type] = handler_type
+    self._handler_to_types[handler_type] = (leaf_type, abstract_type)
 
-  def is_handleable(self, ty: Type[Any]) -> bool:
+  def is_handleable(self, leaf_type: Type[Any]) -> bool:
     """Returns True if the type is handleable."""
-    return self.get(ty) is not None
+    return self._try_get(leaf_type) is not None
 
-  def is_abstract_handleable(self, abstract_ty: Type[Any]) -> bool:
+  def is_abstract_handleable(self, abstract_type: Type[Any]) -> bool:
     """Returns True if the abstract type is handlable."""
-    return self.get_abstract(abstract_ty) is not None
+    return self._try_get_abstract(abstract_type) is not None
 
 
 class StandardLeafHandlerRegistry(BaseLeafHandlerRegistry):
@@ -197,11 +208,10 @@ class StandardLeafHandlerRegistry(BaseLeafHandlerRegistry):
   including jax.Array, np.ndarray, int, float, bytes, and np.number.
   """
 
-  def __init__(self, context: context_lib.Context | None = None):
-    super().__init__(context)
-
+  def __init__(self):
+    super().__init__()
     for (
         ty,
         abstract_ty,
     ), handler_class in STANDARD_TYPE_AND_ABSTRACT_TYPE_TO_HANDLER.items():
-      self.add(ty, abstract_ty, handler_class(context=context))
+      self.add(ty, abstract_ty, handler_class)